Finished tests of Armoured ability
Made the color nullable, so the default color of leather armor is its own color instead of white.

Added manually all the mobs that can wear armor (and it's visible) to the config.yml, so the user can chance or add more if more mobs that can wear armor got released in the future.